Load config from Spatial.toml file
