+++ cargo-spatial/Cargo.toml
@@ -15,3 +15,4 @@ serde_json = "1.0.38"
 spatialos-sdk-code-generator = { path = "../spatialos-sdk-code-generator" }
 structopt = "0.2.14"
 tap = "0.3.1"
+toml = "0.4.10"
